Incorporate cgi.FieldStorage into Webware
Since the cgi module has been deprecated in Python 3.11,
we incorporate the cgi.FieldStorage into the WebUtils package.
This also allows us to make the necessary modifications
and patches directly in the code instead of implementing
them in a subclass which was brittle and broke several times.

This way Webware will keep full backward compatibility.
In a future version, Request.FieldStorage() could be removed
and fields could be provided via the multipart package.